`AddExperience` replaced with `AddSkill` and UI updates
Implemented a new method `AddSkill` in `MainViewModel.cs` to replace the `AddExperience` method. This new method triggers an alert when a new skill is added. The `AddExperience` method, which was responsible for creating a new `ExperienceModel` object, adding it to the `Experiences` collection, and displaying an alert with the serialized data of the new experience, was removed.

In `CreateResumePage.xaml`, a `CollectionView` was added to display the `Experiences` collection. Each item in the collection is displayed in a `HorizontalStackLayout` with a `Label` that binds to the current item.

In `MainPage.xaml`, the `HorizontalOptions` and `VerticalOptions` properties of the `VerticalStackLayout` were modified from `CenterAndExpand` to `Center`.