Refactor code and update dependencies in ResumeBuilderMAUI project
This commit includes several changes to the ResumeBuilderMAUI project. The `ResumeServices` class and its namespace import have been removed from `MauiProgram.cs` and `ResumeServices.cs`, and the singleton service registration for `ResumeServices` has also been removed from `MauiProgram.cs`. The versions of `Microsoft.Maui.Controls` and `Microsoft.Maui.Controls.Compatibility` have been updated from a variable `$(MauiVersion)` to a fixed version `8.0.40` in `ResumeBuilderMAUI.csproj`. A new `Services` folder has been added to the project. In `MainViewModel.cs`, the initializations of `ObservableRangeCollection` for `Experiences`, `Educations`, `Projects`, and `Errors` have been changed to an empty array `[]`, and the same change has been made for the `ObservableCollection` initialization for `skillList`. Lastly, the `Educations` property has been added to the list of properties to be cleared in the `ClearAllEntries` method.