Added nullable Title property and refactored MainViewModel
In the `Experience` class within `Resume.cs`, a new nullable property `Title` of type string has been added. In `MainViewModel.cs`, the creation of a new `Experience` object has been removed and instead, the `experienceData` object is directly added to the `Experiences` collection. The error check and alert display in the `Save` method have been removed. A new `ResumeId` is now generated using the `Generators.RandomNumber()` method when creating a new `Person` object. These changes are part of a refactoring process to simplify the code and improve the data structure.