Added new features and updated UI elements
Added `ResumeDate` property to `Person` class in `Resume.cs` and set it to the current date in `MainViewModel.cs` when a new `Person` object is created. Two new methods, `GetAllPersons` and `GetPersonsByName`, have been added to `LocalDbService.cs` for retrieving persons from the database. Updated "Continue" and "Create New Resume" buttons with `x:Name` properties in `MainPage.xaml` for easier reference in code. Modified the constructor for `MainPage` in `MainPage.xaml.cs`.